// src/app/models/viewUniqueSamplePerService.model.ts
export interface ViewUniqueSamplePerService {
    ulid: number;
    sampleNo: number;
    sampleTypeId: number;
    sampleType: string;
    statusId: number;
    status: string;
    statusColor: string;
    colorName: string;
    serviceCount: number;
    encodedDate: Date;
    referralTypeId: number;
    referredBy: string;
    age: number;
    firstName: string;
    lastName: string;
    gender: string;
    isSelected: boolean;
}